V 1.6

+++ src/account/Account.java
@@ -35,7 +35,7 @@ public abstract class Account {
         this.owner = owner;
     }
 
-    public void deposit(int amount) {
+    public void deposit(double amount) {
         if (amount <= 0) {
             throw new IllegalArgumentException("Deposit amount must be positive.");
         }
